requirements.txt
psycopg2-binary
flask
fastapi
uvicorn
pydantic
pyyaml
requests
python-telegram-bot>=20.0
discord.py>=2.0
edge-tts
